requirements.txt
python-telegram-bot==20.6
wikipedia==1.4.0